pyproject.toml
[project]
name = "uxflows-runner"
version = "0.1.0"
description = "Voice/text runner for UX4 v0 specs — Pipecat-based dispatcher with browser audio bridge."
readme = "README.md"
authors = [
    { name = "Tapan Parikh", email = "tparikh@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pipecat-ai[google,silero,webrtc,websocket]>=1.1.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "simpleeval>=1.0.7",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
uxflows-runner = "uxflows_runner.cli:main"

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
